requirements.txt
arrow==0.12.1
attrs==19.3.0
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
docutils==0.15.2
entrypoints==0.3
idna==2.7
importlib-metadata==0.23
ipython-genutils==0.2.0
jeepney==0.4.1
jsonschema==3.2.0
jupyter-core==4.6.1
keyring==19.2.0
more-itertools==7.2.0
nbformat==4.4.0
numpy==1.17.4
pandas==0.23.4
pkginfo==1.5.0.1
plotly==3.1.1
pycparser==2.19
Pygments==2.4.2
pyrsistent==0.15.5
python-dateutil==2.7.3
python-dotenv==0.15.0
pytz==2018.5
readme-renderer==24.0
requests==2.20.0
requests-toolbelt==0.9.1
retrying==1.3.3
SecretStorage==3.1.1
six==1.13.0
tqdm==4.38.0
traitlets==4.3.3
twine==3.0.0
urllib3==1.24.3
webencodings==0.5.1
zipp==0.6.0
